feat: add exception throwing to get and delete methods if requested account does not exist

+++ banking_microservice_test_task/src/main/java/com/banking/banking_microservice_test_task/controller/BankAccountController.java
@@ -1,8 +1,12 @@
 package com.banking.banking_microservice_test_task.controller;
 
 import com.banking.banking_microservice_test_task.entity.BankAccount;
+import com.banking.banking_microservice_test_task.exception_handling.BankAccountIncorrectData;
+import com.banking.banking_microservice_test_task.exception_handling.NoSuchBankAccountException;
 import com.banking.banking_microservice_test_task.service.BankAccountService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.UUID;
@@ -16,6 +20,11 @@ public class BankAccountController {
     @GetMapping("/account/{id}")
     public BankAccount getAccount(@PathVariable UUID id) {
         BankAccount bankAccount = bankAccountService.getAccount(id);
+
+        if (bankAccount == null) {
+            throw new NoSuchBankAccountException("There is no bank account with id " + id + " in database.");
+        }
+
         return bankAccount;
     }
 
@@ -26,9 +35,14 @@ public class BankAccountController {
     }
 
     @DeleteMapping("/account/{id}")
-    public String deleteAccount(@PathVariable UUID id) {
+    public void deleteAccount(@PathVariable UUID id) {
+        BankAccount bankAccount = bankAccountService.getAccount(id);
+
+        if (bankAccount == null) {
+            throw new NoSuchBankAccountException("There is no bank account with id " + id + " in database, so it cannot be deleted.");
+        }
+
         bankAccountService.deleteAccount(id);
-        return "Account " + id + " was deleted.";
     }
 }
